# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InventoryItem {
  id: number;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sku?: string | null;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  minQuantity?: number | null;
  /** @nullable */
  unitPrice?: number | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  lastUpdated: string;
  createdAt: string;
}

export interface InventoryStats {
  totalItems: number;
  totalValue: number;
  lowStockCount: number;
  categories: number;
  recentlyAdded: number;
  recentlyConsumed: number;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  added: "added",
  consumed: "consumed",
  updated: "updated",
} as const;

export interface Transaction {
  id: number;
  itemId: number;
  itemName: string;
  type: TransactionType;
  quantityChange: number;
  previousQuantity: number;
  newQuantity: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface UploadResult {
  success: boolean;
  message: string;
  itemsProcessed: number;
  itemsAdded: number;
  itemsUpdated: number;
  itemsRemoved: number;
  errors: string[];
  syncedAt: string;
}

export interface LastSync {
  /** @nullable */
  syncedAt: string | null;
  /** @nullable */
  fileName: string | null;
  /** @nullable */
  itemsProcessed: number | null;
}

export type ListInventoryItemsParams = {
  category?: string;
  search?: string;
  lowStock?: string;
};