# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItem {
  id: number;
  name: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  sku?: string | null;
  quantity: number;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  minQuantity?: number | null;
  /** @nullable */
  unitPrice?: number | null;
  /** @nullable */
  supplier?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  description?: string | null;
  lastUpdated: string;
  createdAt: string;
}