# lib/api-zod/src/generated/types/inventoryStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryStats {
  totalItems: number;
  totalValue: number;
  lowStockCount: number;
  categories: number;
  recentlyAdded: number;
  recentlyConsumed: number;
}