# lib/api-zod/src/generated/types/lastSync.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */

export interface LastSync {
  /** @nullable */
  syncedAt: string | null;
  /** @nullable */
  fileName: string | null;
  /** @nullable */
  itemsProcessed: number | null;
}