# lib/api-zod/src/generated/types/transaction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionType } from "./transactionType";

export interface Transaction {
  id: number;
  itemId: number;
  itemName: string;
  type: TransactionType;
  quantityChange: number;
  previousQuantity: number;
  newQuantity: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}